move naming rules to naming package
