+++ detekt-rules/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/providers/NamingProvider.kt
@@ -3,9 +3,9 @@ package io.gitlab.arturbosch.detekt.rules.providers
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.RuleSet
 import io.gitlab.arturbosch.detekt.api.RuleSetProvider
-import io.gitlab.arturbosch.detekt.rules.style.MatchingDeclarationName
-import io.gitlab.arturbosch.detekt.rules.style.MemberNameEqualsClassName
-import io.gitlab.arturbosch.detekt.rules.style.naming.NamingRules
+import io.gitlab.arturbosch.detekt.rules.naming.MatchingDeclarationName
+import io.gitlab.arturbosch.detekt.rules.naming.MemberNameEqualsClassName
+import io.gitlab.arturbosch.detekt.rules.naming.NamingRules
 
 /**
  * The naming ruleset contains rules which assert the naming of different parts of the codebase.
